Part-way to intermediate saves. Cluster-indicies looks broken?
