+++ enspara/apps/cluster.py
@@ -17,6 +17,7 @@ from glob import glob
 
 import numpy as np
 import mdtraj as md
+from enspara.cluster.util import *
 
 try:
     # this mpi will get overriden by the enspara mpi module in a few lines.
@@ -131,6 +132,9 @@ def process_command_line(argv):
         "--cluster-iterations", default=None, type=int,
         help="The number of refinement iterations to perform. This is only "
              "relevant to khybrid clustering.")
+    cluster_args.add_argument(
+        "--save_intermediates", default=False, type=bool,
+        help="Save intermediate clustering results when doing khybrid? ")
     cluster_args.add_argument(
         "--init-center-inds", default=None, type=str,
         help="Path to a .npy file that is a list giving the position of "
@@ -282,199 +286,6 @@ def process_command_line(argv):
     return args
 
 
-def expand_files(pgroups):
-    expanded_pgroups = []
-    for pgroup in pgroups:
-        expanded_pgroups.append([])
-        for p in pgroup:
-            expanded_pgroups[-1].extend(sorted(glob(p)))
-    return expanded_pgroups
-
-
-def load_features(features, stride):
-    try:
-        if len(features) == 1:
-            with timed("Loading features took %.1f s.", logger.info):
-                lengths, data = mpi.io.load_h5_as_striped(features[0], stride)
-
-        else:  # and len(features) > 1
-            with timed("Loading features took %.1f s.", logger.info):
-                lengths, data = mpi.io.load_npy_as_striped(features, stride)
-
-        with timed("Turned over array in %.2f min", logger.info):
-            tmp_data = data.copy()
-            del data
-            data = tmp_data
-    except MemoryError:
-        logger.error(
-            "Ran out of memory trying to allocate features array"
-            " from file %s", features[0])
-        raise
-
-    logger.info("Loaded %s trajectories with %s frames with stride %s.",
-                len(lengths), len(data), stride)
-
-    return lengths, data
-
-
-def load_trajectories(topologies, trajectories, selections, stride, processes):
-
-    for top, selection in zip(topologies, selections):
-        sentinel_trj = md.load(top)
-        try:
-            # noop, but causes fast-fail w/bad args.atoms
-            sentinel_trj.top.select(selection)
-        except:
-            raise exception.ImproperlyConfigured((
-                "The provided selection '{s}' didn't match the topology "
-                "file, {t}").format(s=selection, t=top))
-
-    flat_trjs = []
-    configs = []
-    n_inds = None
-
-    for topfile, trjset, selection in zip(topologies, trajectories,
-                                          selections):
-        top = md.load(topfile).top
-        indices = top.select(selection)
-
-        if n_inds is not None:
-            if n_inds != len(indices):
-                raise exception.ImproperlyConfigured(
-                    ("Selection on topology %s selected %s atoms, but "
-                     "other selections selected %s atoms.") %
-                    (topfile, len(indices), n_inds))
-        n_inds = len(indices)
-
-        for trj in trjset:
-            flat_trjs.append(trj)
-            configs.append({
-                'top': top,
-                'stride': stride,
-                'atom_indices': indices,
-            })
-
-    logger.info(
-        "Loading %s trajectories with %s atoms using %s processes "
-        "(subsampling %s)",
-        len(flat_trjs), len(top.select(selection)), processes, stride)
-    assert len(top.select(selection)) > 0, "No atoms selected for clustering"
-
-    with timed("Loading took %.1f sec", logger.info):
-        lengths, xyz = mpi.io.load_trajectory_as_striped(
-            flat_trjs, args=configs, processes=auto_nprocs())
-
-    with timed("Turned over array in %.2f min", logger.info):
-        tmp_xyz = xyz.copy()
-        del xyz
-        xyz = tmp_xyz
-
-    logger.info("Loaded %s frames.", len(xyz))
-
-    return lengths, xyz, top.subset(top.select(selection))
-
-
-def load_asymm_frames(center_indices, trajectories, topology, subsample):
-
-    frames = []
-    begin_index = 0
-    for topfile, trjset in zip(topology, trajectories):
-        end_index = begin_index + len(trjset)
-        target_centers = [c for c in center_indices
-                          if begin_index <= c[0] < end_index]
-
-        try:
-            subframes = load_frames(
-                list(itertools.chain(*trajectories)),
-                target_centers,
-                top=md.load(topfile).top,
-                stride=subsample)
-        except exception.ImproperlyConfigured:
-            logger.error('Failure to load cluster centers %s for topology %s',
-                         topfile, target_centers)
-            raise
-
-        frames.extend(subframes)
-        begin_index += len(trjset)
-
-    return frames
-
-
-def load_trjs_or_features(args):
-
-    if args.features:
-        with timed("Loading features took %.1f s.", logger.info):
-            lengths, data = load_features(args.features, stride=args.subsample)
-    else:
-        assert args.trajectories
-        assert len(args.trajectories) == len(args.topologies)
-
-        targets = {os.path.basename(topf): "%s files" % len(trjfs)
-                   for topf, trjfs
-                   in zip(args.topologies, args.trajectories)
-                   }
-        logger.info("Beginning clustering; targets:\n%s",
-                    json.dumps(targets, indent=4))
-
-        with timed("Loading trajectories took %.1f s.", logger.info):
-            lengths, xyz, select_top = load_trajectories(
-                args.topologies, args.trajectories, selections=args.atoms,
-                stride=args.subsample, processes=auto_nprocs())
-
-        logger.info("Clustering using %s atoms matching '%s'.", xyz.shape[1],
-                    args.atoms)
-
-        # md.rmsd requires an md.Trajectory object, so wrap `xyz` in
-        # the topology.
-        data = md.Trajectory(xyz=xyz, topology=select_top)
-
-    return lengths, data
-
-
-def write_centers_indices(path, indices):
-    if path:
-        with open(path, 'wb') as f:
-            np.save(f, indices)
-    else:
-        logger.info("--center-indices not provided, not writing center "
-                    "indices to file.")
-
-
-def write_centers(result, args):
-    if args.features:
-        np.save(args.center_features, result.centers)
-    else:
-        outdir = os.path.dirname(args.center_features)
-        logger.info("Saving cluster centers at %s", outdir)
-
-        try:
-            os.makedirs(outdir)
-        except FileExistsError:
-            pass
-
-        centers = load_asymm_frames(result.center_indices, args.trajectories,
-                                    args.topologies, args.subsample)
-        with open(args.center_features, 'wb') as f:
-            pickle.dump(centers, f)
-
-
-def write_assignments_and_distances_with_reassign(result, args):
-    if args.subsample == 1:
-        logger.debug("Subsampling was 1, not reassigning.")
-        ra.save(args.distances, result.distances)
-        ra.save(args.assignments, result.assignments)
-    elif not args.no_reassign:
-        logger.debug("Reassigning data from subsampling of %s", args.subsample)
-        assig, dist = reassign(
-            args.topologies, args.trajectories, args.atoms,
-            centers=result.centers)
-
-        ra.save(args.distances, dist)
-        ra.save(args.assignments, assig)
-    else:
-        logger.debug("Got --no-reassign, not doing reassigment")
-
-
 def main(argv=None):
 
     args = process_command_line(argv)
@@ -489,6 +300,8 @@ def main(argv=None):
             kwargs['kmedoids_updates'] = int(args.cluster_iterations)
         elif args.Clusterer is KMedoids:
             kwargs['n_iters'] = int(args.cluster_iterations)
+        if args.Clusterer is not KCenters:
+            kwargs['args']=args
 
     #kmedoids doesn't need a cluster radius, but kcenters does
     if args.cluster_radius is not None:
@@ -509,6 +322,8 @@ def main(argv=None):
         if args.init_assignments:
             kwargs_restart['X_lengths'], kwargs_restart['assignments'] = \
                 mpi.io.load_h5_as_striped(args.init_assignments)
+        if args.save_intermediates:
+            kwargs_restart['args']=args
         if args.init_center_inds:
             kwargs_restart['cluster_center_inds'] = \
                 np.load(args.init_center_inds)
